tests/wpt/tests/html/canvas/element/path-objects/2d.path.stroke.union.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.stroke.union</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.stroke.union</h1>
<p class="desc">Strokes in opposite directions are unioned, not subtracted</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Strokes in opposite directions are unioned, not subtracted");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 40;
  ctx.moveTo(0, 10);
  ctx.lineTo(100, 10);
  ctx.moveTo(100, 40);
  ctx.lineTo(0, 40);
  ctx.stroke();

  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
